+++ Changes.md
@@ -53,6 +53,7 @@ Improvements
 - HierarchyView :
   - Added support for inclusion and exclusion of leaf level locations to the Visible Set.
   - Added <kbd>F</kbd> shortcut to expand ancestors and frame selected paths. This replaces the previous behaviour of automatic updates to expansion and framing in the HierarchyView on selection change.
+- CyclesOptions : Added `useLightTree` plug.
 
 Fixes
 -----